Update the cleaning request flow
Change the cleaning service tests and request.controller

Apply the OneHelpSDK to the proxy requests controller and tests

+++ proxy/app/Http/Controllers/RequestsController.php
@@ -7,11 +7,10 @@ use Exception;
 use App\User;
 use App\Request;
 
-use Illuminate\Support\Str;
 use Illuminate\Http\Request as HttpRequest;
 use Illuminate\Http\Response;
 use Illuminate\Http\JsonResponse;
-
+use OneHelpSDK\Facades\Cleaning;
 class RequestsController extends Controller
 {
     public function list(): JsonResponse
@@ -23,6 +22,10 @@ class RequestsController extends Controller
     {
         try {
             $data = $http->all();
+            $requestApi = Cleaning::createRequest($data);
+            if (Cleaning::getStatus() !== 201) {
+                throw new Exception('An error occurred when the request was saved');
+            }
             $user = User::firstOrCreate($data['user']);
             if (empty($user->id)) {
                 $user->save();
@@ -30,7 +33,7 @@ class RequestsController extends Controller
             $data['user_id'] = $user->id;
             unset($data['user']);
             $request = new Request($data);
-            $request->reference = Str::uuid();
+            $request->reference = $requestApi->_id;
             $request->save();
             return response()->json(null, Response::HTTP_CREATED);
         } catch (Exception $e) {
@@ -42,7 +45,11 @@ class RequestsController extends Controller
     {
         try {
             $data = $http->all();
-            $request = Request::firstOrFail($id);
+            $request = Request::findOrFail($id);
+            Cleaning::updateRequest($request->reference, $data);
+            if (Cleaning::getStatus() !== 201) {
+                throw new Exception('An error occurred when the request was updated');
+            }
             $this->fillWithData($request, $data);
             $request->save();
             return response()->json(null, Response::HTTP_NO_CONTENT);
@@ -71,7 +78,11 @@ class RequestsController extends Controller
     public function delete(int $id): JsonResponse
     {
         try {
-            $request = Request::firstOrFail($id);
+            $request = Request::findOrFail($id);
+            Cleaning::deleteRequest($request->reference);
+            if (Cleaning::getStatus() !== 204) {
+                throw new Exception('An error occurred when the request was updated');
+            }
             $request->delete();
             return response()->json(null, Response::HTTP_NO_CONTENT);
         } catch (Exception $e) {
